pyproject.toml
[project]
name = "play-lmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jon Goikoetxea", email = "jon.goikoetxea.m@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "beartype>=0.21.0",
    "equinox>=0.13.0",
    "gymnasium-robotics>=1.4.1",
    "hydra-core>=1.3.2",
    "jax[cuda12]>=0.7.0",
    "jaxtyping>=0.3.2",
    "jmp>=0.0.4",
    "language-table",
    "minari>=0.5.3",
    "optax>=0.2.5",
    "tf-nightly<2.21",
    "tqdm>=4.67.1",
]

[project.scripts]
play-lmp = "play_lmp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
language-table = { git = "https://github.com/google-research/language-table" }

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]

[tool.ruff]
lint.ignore = ["F722"]
